Clean up logging

+++ tunnel.h
@@ -1,13 +1,14 @@
 #ifndef __TUNNEL_H
 #define __TUNNEL_H
 
+#include <sys/socket.h>
 #include <net/if.h>
 
 struct gre_host;
 
 struct tunnel{
     char 		     name[IFNAMSIZ];
-    int              id;
+    unsigned short   id;
 	struct gre_host* dest;
 
     int              tun_fd;
@@ -29,7 +30,7 @@ struct tunnel* tunnel_alloc();
 int tunnel_compar(const void* _t1, const void* _t2);
 
 /* Create a new tunnel with the provided data */
-struct tunnel* tunnel_new(char* name, int tunnel_id,
+struct tunnel* tunnel_new(char* name, unsigned short tunnel_id,
 						  struct gre_host* host);
 
 /* Opens and configures the tunnel device */
